Se reemplaza el double por int

+++ CUDA/00_cuda_simple_comparacion.cu
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 
-__global__ void sumParallel(double *dev_sum, double num)
+__global__ void sumParallel(int *dev_sum, int num)
 {
     int tid = threadIdx.x + blockIdx.x * blockDim.x;
     int stride = blockDim.x * gridDim.x;
@@ -10,40 +10,40 @@ __global__ void sumParallel(double *dev_sum, double num)
     {
 
 	/* se usa la función atomicAdd para realizar la sumatoria y  garantizar que varios hilos no vayan a escribir simultáneamente en la misma ubicación de memoria. */
-        atomicAdd(dev_sum, (double)tid);
+        atomicAdd(dev_sum, (int)tid);
         tid += stride; //coge el identificado unico de cada hilo y le agrega el paso para pasar al siguiente hilo
     }
 }
 
-double cuda_suma_simple(double num)
+int cuda_suma_simple(int num)
 {
     //int num, sum = 0;
-    double sum = 0;
+    int sum = 0;
 
-    double *dev_sum; //referencia de la variable que se ubicara en el device
+    int *dev_sum; //referencia de la variable que se ubicara en el device
 
   
-    cudaMalloc((void**)&dev_sum, sizeof(double)); //reservamos espacio de memoria
-    cudaMemcpy(dev_sum, &sum, sizeof(double), cudaMemcpyHostToDevice); //copaimos la variable desde el host al sum
+    cudaMalloc((void**)&dev_sum, sizeof(int)); //reservamos espacio de memoria
+    cudaMemcpy(dev_sum, &sum, sizeof(int), cudaMemcpyHostToDevice); //copaimos la variable desde el host al sum
 
     int blockSize = 256;
     int gridSize = (num + blockSize - 1) / blockSize;
 
     sumParallel<<<gridSize, blockSize>>>(dev_sum, num); //invocamos el kernel sumParallel que es el que se encarga de realizar la suma
     
-    cudaMemcpy(&sum, dev_sum, sizeof(double), cudaMemcpyDeviceToHost); //copiamos el resultado ahora en sentido contrario, es decir desde el device hasta el host
+    cudaMemcpy(&sum, dev_sum, sizeof(int), cudaMemcpyDeviceToHost); //copiamos el resultado ahora en sentido contrario, es decir desde el device hasta el host
     cudaFree(dev_sum); //liberamos la memoria reservada
 
     return sum;
 }
-double sec_suma_simple(double num){
-    double resultado = 0;
+int sec_suma_simple(int num){
+    int resultado = 0;
     for(int i=0; i<=num; i++){
          resultado+=i;
     }
     return resultado;
 }
-float run_cuda_suma_simple(double num){
+float run_cuda_suma_simple(int num){
 
      /*Inicializamos las variables con las cuales tomaremos el tiempo */
     cudaEvent_t start, stop;
@@ -54,7 +54,7 @@ float run_cuda_suma_simple(double num){
     cudaEventCreate(&stop);
     cudaEventRecord(start, 0); //comienza a tomar el tiempo
 
-    double result = cuda_suma_simple(num);
+    int result = cuda_suma_simple(num);
     printf("Resultado Cuda = %f\n", result);
 
     cudaEventRecord(stop, 0); //para de tomar el tiempo
@@ -66,7 +66,7 @@ float run_cuda_suma_simple(double num){
     return elapsedTime;
 }
 
-float run_sec_suma_simple(double num){
+float run_sec_suma_simple(int num){
  
      /*Inicializamos las variables con las cuales tomaremos el tiempo */
     cudaEvent_t start, stop;
@@ -78,7 +78,7 @@ float run_sec_suma_simple(double num){
     cudaEventRecord(start, 0); //comienza a tomar el tiempo
 
 
-    double result = sec_suma_simple(num);
+    int result = sec_suma_simple(num);
     printf("Resultado Secu = %f\n", result);
 
     cudaEventRecord(stop, 0); //para de tomar el tiempo
